changed emission score to prob conversion

+++ src/transitionprobabilitycomputer.cpp
@@ -8,7 +8,7 @@
 using namespace std;
 
 TransitionProbabilityComputer::TransitionProbabilityComputer(const float& recombcost, const vector<int>& next_allele_reference) {
-    int s = next_allele_reference.size();
+    double s = (double)next_allele_reference.size();
     float r = recombcost;
     this->pr = (1.0 - exp(-(r/s))) / s;
     this->qr = exp(-(r/s)) + pr;
